+++ tests/testthat/test_attribute_size.R
@@ -27,3 +27,33 @@ test_that(
     expect_equal(N, 4050)
   }
 )
+
+test_that(
+  "new workflow for size of p-attribute",
+  {
+    conventional <- cl_attribute_size(
+      corpus = "REUTERS",
+      attribute = "word",
+      attribute_type = "p",
+      registry = get_tmp_registry()
+    )
+    attr <- p_attr("REUTERS", p_attribute = "word", registry = get_tmp_registry())
+    new <- p_attr_size(attr)
+    expect_identical(conventional, new)
+  }
+)
+
+test_that(
+  "new workflow for size of s-attribute",
+  {
+    conventional <- cl_attribute_size(
+      corpus = "REUTERS",
+      attribute = "id",
+      attribute_type = "s",
+      registry = get_tmp_registry()
+    )
+    attr <- s_attr("REUTERS", s_attribute = "id", registry = get_tmp_registry())
+    new <- s_attr_size(attr)
+    expect_identical(conventional, new)
+  }
+)